+++ src/Phan/Config.php
@@ -537,6 +537,12 @@ class Config
         'read_magic_method_annotations' => true,
 
         // If disabled, Phan will not read docblock type
+        // annotation comments for `@mixin`.
+        //
+        // Note: `read_type_annotations` must also be enabled.
+        'read_mixin_annotations' => true,
+
+        // If disabled, Phan will not read docblock type
         // annotation comments (such as for `@return`, `@param`,
         // `@var`, `@suppress`, `@deprecated`) and only rely on
         // types expressed in code.